Rename env params to static params
