+++ core/schains/config/helper.py
@@ -30,7 +30,7 @@ from skale.dataclasses.skaled_ports import SkaledPorts
 from core.schains.config.directory import schain_config_filepath
 from core.schains.dkg.utils import get_secret_key_share_filepath
 from tools.helper import read_json
-from tools.configs import ENVIRONMENT_PARAMS_FILEPATH, ENV_TYPE
+from tools.configs import STATIC_PARAMS_FILEPATH, ENV_TYPE
 from tools.configs.containers import LOCAL_IP
 from tools.helper import safe_load_yml
 
@@ -38,7 +38,7 @@ from tools.helper import safe_load_yml
 logger = logging.getLogger(__name__)
 
 
-def get_environment_params(env_type=ENV_TYPE, path=ENVIRONMENT_PARAMS_FILEPATH):
+def get_static_params(env_type=ENV_TYPE, path=STATIC_PARAMS_FILEPATH):
     ydata = safe_load_yml(path)
     return ydata['envs'][env_type]
