+++ tests/skale-data/config/static_params.yaml
