+++ tools/configs/__init__.py
@@ -74,7 +74,7 @@ ALLOWED_TIMESTAMP_DIFF = int(os.getenv('ALLOWED_TIMESTAMP_DIFF', 120))
 ENV_TYPE = os.environ.get('ENV_TYPE')
 ALLOCATION_FILEPATH = os.path.join(CONFIG_FOLDER, 'schain_allocation.yml')
 
-ENVIRONMENT_PARAMS_FILEPATH = os.path.join(CONFIG_FOLDER, 'environment_params.yaml')
+STATIC_PARAMS_FILEPATH = os.path.join(CONFIG_FOLDER, 'static_params.yaml')
 
 DEFAULT_POOL = 'transactions'
