+++ tools/docker_utils.py
@@ -251,9 +251,8 @@ class DockerUtils:
         self,
         container_name: Container,
         head: int = 100,
-        tail: int = 200,
-        to_loggger: bool = True
-    ) -> None:
+        tail: int = 200
+    ) -> str:
         container = self.safe_get_container(container_name)
         if not container:
             return
@@ -264,10 +263,8 @@ class DockerUtils:
         )
         pretext = f'container {container_name} logs: \n'
         logs = (head_lines + CONTAINER_LOGS_SEPARATOR + tail_lines).decode("utf-8")
-        if to_loggger:
-            logger.info(pretext + logs)
-        else:
-            print(pretext + logs)
+        logger.info(pretext + logs)
+        return logs
 
     @classmethod
     def save_container_logs(
